requirements3.txt
# requirements.txt  – append
torch>=2.2            # CUDA wheels auto-detect GPU env
torchmetrics          # nice MAE/RMSE helpers
tqdm
optuna                # lightweight Bayesian search (optional)
pandas-stubs>=2.1        # typing alias restores SequenceNotStr for pandas 2.2.x
